Frontend/fix auth

+++ frontend/src/services/hairmodel.js
@@ -8,10 +8,27 @@ const getAll = async () => {
     const response = await axios.get(baseUrl)
     return response.data
   } catch (err) {
-    if (err.response.status === 403) {
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'Authentication credentials were not provided.'
+      )
+    ) {
       window.localStorage.clear()
       logoutService.logout()
       return { error: true, status: err.response.status }
+    }
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'You do not have permission to perform this action.'
+      )
+    ) {
+      return {
+        error: true,
+        status: err.response.status,
+        detail: err.response.data.detail,
+      }
     }
   }
 }
@@ -21,10 +38,27 @@ const getOne = async (id) => {
     const response = await axios.get(`${baseUrl}${id}/`)
     return response.data
   } catch (err) {
-    if (err.response.status === 403) {
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'Authentication credentials were not provided.'
+      )
+    ) {
       window.localStorage.clear()
       logoutService.logout()
       return { error: true, status: err.response.status }
+    }
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'You do not have permission to perform this action.'
+      )
+    ) {
+      return {
+        error: true,
+        status: err.response.status,
+        detail: err.response.data.detail,
+      }
     }
   }
 }
@@ -39,9 +73,26 @@ const update = async (updatedHairModel, id) => {
     const response = await axios.put(`${baseUrl}${id}/`, updatedHairModel)
     return response.data
   } catch (err) {
-    if (err.response.status === 403) {
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'Authentication credentials were not provided.'
+      )
+    ) {
       window.localStorage.clear()
       logoutService.logout()
+    }
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'You do not have permission to perform this action.'
+      )
+    ) {
+      return {
+        error: true,
+        status: err.response.status,
+        detail: err.response.data.detail,
+      }
     }
   }
 }
@@ -51,9 +102,26 @@ const del = async (id) => {
     const response = await axios.delete(`${baseUrl}${id}/`)
     return response.data
   } catch (err) {
-    if (err.response.status === 403) {
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'Authentication credentials were not provided.'
+      )
+    ) {
       window.localStorage.clear()
       logoutService.logout()
+    }
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'You do not have permission to perform this action.'
+      )
+    ) {
+      return {
+        error: true,
+        status: err.response.status,
+        detail: err.response.data.detail,
+      }
     }
   }
 }
