+++ frontend/src/services/service.js
@@ -13,10 +13,27 @@ const getAll = async () => {
     const response = await axios.get(serviceUrl)
     return response.data
   } catch (err) {
-    if (err.response.status === 403) {
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'Authentication credentials were not provided.'
+      )
+    ) {
       window.localStorage.clear()
       logoutService.logout()
       return { error: true, status: err.response.status }
+    }
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'You do not have permission to perform this action.'
+      )
+    ) {
+      return {
+        error: true,
+        status: err.response.status,
+        detail: err.response.data.detail,
+      }
     }
   }
 }
@@ -26,10 +43,27 @@ const getOne = async (id) => {
     const response = await axios.get(`${serviceUrl}${id}/`)
     return response.data
   } catch (err) {
-    if (err.response.status === 403) {
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'Authentication credentials were not provided.'
+      )
+    ) {
       window.localStorage.clear()
       logoutService.logout()
       return { error: true, status: err.response.status }
+    }
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'You do not have permission to perform this action.'
+      )
+    ) {
+      return {
+        error: true,
+        status: err.response.status,
+        detail: err.response.data.detail,
+      }
     }
   }
 }
@@ -39,10 +73,27 @@ const create = async (newService) => {
     const response = await axios.post(serviceUrl, newService)
     return response.data
   } catch (err) {
-    if (err.response.status === 403) {
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'Authentication credentials were not provided.'
+      )
+    ) {
       window.localStorage.clear()
       logoutService.logout()
       return { error: true, status: err.response.status }
+    }
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'You do not have permission to perform this action.'
+      )
+    ) {
+      return {
+        error: true,
+        status: err.response.status,
+        detail: err.response.data.detail,
+      }
     }
   }
 }
@@ -52,10 +103,27 @@ const update = async (id, updatedService) => {
     const response = await axios.put(`${serviceUrl}${id}/`, updatedService)
     return response.data
   } catch (err) {
-    if (err.response.status === 403) {
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'Authentication credentials were not provided.'
+      )
+    ) {
       window.localStorage.clear()
       logoutService.logout()
       return { error: true, status: err.response.status }
+    }
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'You do not have permission to perform this action.'
+      )
+    ) {
+      return {
+        error: true,
+        status: err.response.status,
+        detail: err.response.data.detail,
+      }
     }
   }
 }
@@ -65,10 +133,27 @@ const del = async (id) => {
     const response = await axios.delete(`${serviceUrl}${id}/`)
     return response.data
   } catch (err) {
-    if (err.response.status === 403) {
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'Authentication credentials were not provided.'
+      )
+    ) {
       window.localStorage.clear()
       logoutService.logout()
       return { error: true, status: err.response.status }
+    }
+    if (
+      err.response.status === 403 &&
+      err.response.data.detail.includes(
+        'You do not have permission to perform this action.'
+      )
+    ) {
+      return {
+        error: true,
+        status: err.response.status,
+        detail: err.response.data.detail,
+      }
     }
   }
 }
